requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0

# --------- loggers --------- #
wandb==0.15.0

# --------- huggingface --------- #
timm==0.9.2

# --------- pillow --------- #
pillow==9.4.0

# --------- others --------- #
scipy==1.8.1
pandas==1.5.3
tensorboard==2.11.2
pyrootutils==1.0.4
tqdm==4.65.0
